backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
gunicorn
requests
Werkzeug

tensorflow
scikit-learn
Pillow